Made DatabaseHelper public
